Move auxiliary graph methods that depend on the graph to the graph class as private methods. With this the general _utils script was deleted
