+++ pywgraph/graphs/__init__.py
@@ -1,4 +1,5 @@
 from ._edge import *
 from ._graph import *
+from ._cycles import *
 
 __all__ = [s for s in dir() if not s.startswith("_")]
